using mainTest.cpp for unit tests
