+++ .gitignore
@@ -4,21 +4,49 @@
 *.dll
 *.so
 *.dylib
+*.elf
+*.app
 
 # Ignore compiled files
 *.o
 *.obj
 
+# Ignore static libraries
+*.a
+*.lib
+
+# Ignore shared libraries
+*.so
+*.dylib
+*.dll
+
 # Ignore files generated by IDEs
 .vscode/
 .vscode/*
+.idea/
+*.sublime-project
+*.sublime-workspace
 
 # Ignore temporary files
 tempCodeRunnerFile.cpp
+*.log
+*.tmp
+*.swp
+*~
 
-# Ignore user-generated files
-UT_users_list
-users-list
-UT_user
+# Ignore core dumps
+core
+*.core
 
+# Ignore build directories
+/build/
+/bin/
+/obj/
+/lib/
 
+# Ignore dependency directories
+/node_modules/
+/vendor/
+
+# Ignore files without extension
+*
